Correccion de errores

+++ Vistas/clientes_acciones.php
@@ -1,14 +1,14 @@
 <?php
     include("../Controlador/ClientesControl.php");
     $clientes_control = new ClientesControl();
+    
 
-    if(isset($_POST)){
+    if(isset($_POST) && array_key_exists('id_cliente', $_POST)){
         $data = $_POST;
         if ($data['id_cliente']!= null && $data['id_cliente'] != '' && !isset($data['del'])){
             $clientes_control->update($data);
         } else{
             if($data['id_cliente']!= null && $data['id_cliente'] != '' && isset($data['del']) && $data['del']=='si'){
-                $clientes_control->delete($data);
                 if($data['estado'] == 1){
                     $clientes_control->delete($data);
                 }else{
@@ -25,6 +25,8 @@
         } 
         
     }
+
+    //echo "<br><br><br><a href='Cliente.php'>Clientes</a>"
 
     header("Location: Cliente.php");
     die();
